Extract HelloWorld protocol into separate test module
