+++ xtra-libp2p/tests/basic.rs
@@ -1,9 +1,5 @@
-use anyhow::Context as _;
-use anyhow::Result;
-use async_trait::async_trait;
-use asynchronous_codec::Bytes;
-use futures::SinkExt;
-use futures::StreamExt;
+use crate::hello_world::hello_world_dialer;
+use crate::hello_world::HelloWorld;
 use libp2p_core::multiaddr::Protocol;
 use libp2p_core::Multiaddr;
 use std::collections::HashSet;
@@ -14,15 +10,14 @@ use util::Node;
 use xtra::message_channel::MessageChannel;
 use xtra::spawn::TokioGlobalSpawnExt;
 use xtra::Actor;
-use xtra::Context;
 use xtra_libp2p::Connect;
 use xtra_libp2p::Disconnect;
 use xtra_libp2p::GetConnectionStats;
 use xtra_libp2p::ListenOn;
 use xtra_libp2p::NewInboundSubstream;
 use xtra_libp2p::OpenSubstream;
-use xtra_productivity::xtra_productivity;
 
+mod hello_world;
 mod util;
 
 #[tokio::test]
@@ -30,7 +25,7 @@ async fn hello_world() {
     let alice_hello_world_handler = HelloWorld::default().create(None).spawn_global();
     let (alice, bob, _) = alice_and_bob(
         [(
-            "/hello-world/1.0.0",
+            hello_world::PROTOCOL_NAME,
             alice_hello_world_handler.clone().into(),
         )],
         [],
@@ -41,7 +36,7 @@ async fn hello_world() {
         .endpoint
         .send(OpenSubstream::single_protocol(
             alice.peer_id,
-            "/hello-world/1.0.0",
+            hello_world::PROTOCOL_NAME,
         ))
         .await
         .unwrap()
@@ -244,7 +239,7 @@ async fn chooses_first_protocol_in_list_of_multiple() {
     let alice_hello_world_handler = HelloWorld::default().create(None).spawn_global();
     let (alice, bob, _) = alice_and_bob(
         [(
-            "/hello-world/1.0.0",
+            hello_world::PROTOCOL_NAME,
             alice_hello_world_handler.clone().into(),
         )],
         [],
@@ -256,7 +251,7 @@ async fn chooses_first_protocol_in_list_of_multiple() {
         .send(OpenSubstream::multiple_protocols(
             alice.peer_id,
             vec![
-                "/hello-world/1.0.0",
+                hello_world::PROTOCOL_NAME,
                 "/foo-bar/1.0.0", // This is unsupported by Alice.
             ],
         ))
@@ -266,7 +261,7 @@ async fn chooses_first_protocol_in_list_of_multiple() {
         .await
         .unwrap();
 
-    assert_eq!(actual_protocol, "/hello-world/1.0.0");
+    assert_eq!(actual_protocol, hello_world::PROTOCOL_NAME);
 }
 
 #[cfg_attr(debug_assertions, tokio::test)] // The assertion for duplicate handlers only runs in debug mode.
@@ -275,8 +270,11 @@ async fn disallow_duplicate_handlers() {
     let hello_world_handler = HelloWorld::default().create(None).spawn_global();
 
     make_node([
-        ("/hello-world/1.0.0", hello_world_handler.clone().into()),
-        ("/hello-world/1.0.0", hello_world_handler.into()),
+        (
+            hello_world::PROTOCOL_NAME,
+            hello_world_handler.clone().into(),
+        ),
+        (hello_world::PROTOCOL_NAME, hello_world_handler.into()),
     ]);
 }
 
@@ -285,7 +283,7 @@ async fn falls_back_to_next_protocol_if_unsupported() {
     let alice_hello_world_handler = HelloWorld::default().create(None).spawn_global();
     let (alice, bob, _) = alice_and_bob(
         [(
-            "/hello-world/1.0.0",
+            hello_world::PROTOCOL_NAME,
             alice_hello_world_handler.clone().into(),
         )],
         [],
@@ -298,7 +296,7 @@ async fn falls_back_to_next_protocol_if_unsupported() {
             alice.peer_id,
             vec![
                 "/foo-bar/1.0.0", // This is unsupported by Alice.
-                "/hello-world/1.0.0",
+                hello_world::PROTOCOL_NAME,
             ],
         ))
         .await
@@ -307,7 +305,7 @@ async fn falls_back_to_next_protocol_if_unsupported() {
         .await
         .unwrap();
 
-    assert_eq!(actual_protocol, "/hello-world/1.0.0");
+    assert_eq!(actual_protocol, hello_world::PROTOCOL_NAME);
 }
 
 async fn alice_and_bob<const AN: usize, const BN: usize>(
@@ -339,51 +337,4 @@ async fn alice_and_bob<const AN: usize, const BN: usize>(
         .unwrap();
 
     (alice, bob, alice_listen)
-}
-
-#[derive(Default)]
-struct HelloWorld;
-
-#[xtra_productivity]
-impl HelloWorld {
-    async fn handle(&mut self, msg: NewInboundSubstream, ctx: &mut Context<Self>) {
-        tracing::info!("New hello world stream from {}", msg.peer);
-
-        tokio_extras::spawn_fallible(
-            &ctx.address().unwrap(),
-            hello_world_listener(msg.stream),
-            move |e| async move {
-                tracing::warn!("Hello world protocol with peer {} failed: {}", msg.peer, e);
-            },
-        );
-    }
-}
-
-#[async_trait]
-impl Actor for HelloWorld {
-    type Stop = ();
-
-    async fn stopped(self) -> Self::Stop {}
-}
-
-async fn hello_world_dialer(stream: xtra_libp2p::Substream, name: &'static str) -> Result<String> {
-    let mut stream = asynchronous_codec::Framed::new(stream, asynchronous_codec::LengthCodec);
-
-    stream.send(Bytes::from(name)).await?;
-    let bytes = stream.next().await.context("Expected message")??;
-    let message = String::from_utf8(bytes.to_vec())?;
-
-    Ok(message)
-}
-
-async fn hello_world_listener(stream: xtra_libp2p::Substream) -> Result<()> {
-    let mut stream =
-        asynchronous_codec::Framed::new(stream, asynchronous_codec::LengthCodec).fuse();
-
-    let bytes = stream.select_next_some().await?;
-    let name = String::from_utf8(bytes.to_vec())?;
-
-    stream.send(Bytes::from(format!("Hello {name}!"))).await?;
-
-    Ok(())
 }
